src/main/java/com/chy/smdc/service/impl/ProductCategoryServiceImpl.java
package com.chy.smdc.service.impl;

import com.chy.smdc.bean.ProductCategory;
import com.chy.smdc.dao.ProductCategoryDao;
import com.chy.smdc.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {
    @Autowired
    ProductCategoryDao productCategoryDao;
    @Override
    public ProductCategory findOne(Integer categoryId) {
        return productCategoryDao.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryDao.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTyprList) {
        return productCategoryDao.findByCategoryTypeIn(categoryTyprList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryDao.save(productCategory);
    }
}
